Implement form save, load data

+++ app/src/main/java/caresurvey/sci/com/caresurvey/activity/DisplayAll_Activity.java
@@ -67,7 +67,7 @@ public class DisplayAll_Activity extends AppCompatActivity {
             int k=0;
             int f= formItems.size();
 
-            int[] id_admin=new int[f];
+            long[] id_admin=new long[f];
             String[] name_admin=new String[f];
             final int[] status_admin= new int[f];
             final String[] inS= new String[f];
@@ -168,7 +168,7 @@ public class DisplayAll_Activity extends AppCompatActivity {
             int k = 0;
             int f = formItems.size();
 
-            int[] id_admin = new int[f];
+            long[] id_admin = new long[f];
             String[] name_admin = new String[f];
             final int[] status_admin = new int[f];
             final String[] inS = new String[f];
